third_party/WebKit/LayoutTests/http/tests/inspector/security/mixed-form.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../security-test.js"></script>
<script>
function test()
{
    function dumpExplanations() {
      var explanations = Security.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-explanation");
      for (var i = 0; i < explanations.length; i++)
          InspectorTest.dumpDeepInnerHTML(explanations[i]);
    }

    /** @type {!Protocol.Security.InsecureContentStatus} */
    var insecureContentStatus = { ranMixedContent: false, displayedMixedContent: false, containedMixedForm: true, ranContentWithCertErrors: false, displayedContentWithCertErrors: false, ranInsecureContentStyle: Protocol.Security.SecurityState.Insecure, displayedInsecureContentStyle: Protocol.Security.SecurityState.Neutral };

    InspectorTest.mainTarget.model(Security.SecurityModel).dispatchEventToListeners(Security.SecurityModel.Events.SecurityStateChanged, new Security.PageSecurityState(Protocol.Security.SecurityState.Neutral, true, [], insecureContentStatus, null));

    InspectorTest.addResult("\nSecurity Explanations (Mixed Form Only) --------------");
    dumpExplanations();

    // Now simulate a state change that adds passive mixed content.
    insecureContentStatus = { ranMixedContent: false, displayedMixedContent: true, containedMixedForm: true, ranContentWithCertErrors: false, displayedContentWithCertErrors: false, ranInsecureContentStyle: Protocol.Security.SecurityState.Insecure, displayedInsecureContentStyle: Protocol.Security.SecurityState.Neutral };
    InspectorTest.mainTarget.model(Security.SecurityModel).dispatchEventToListeners(Security.SecurityModel.Events.SecurityStateChanged, new Security.PageSecurityState(Protocol.Security.SecurityState.Neutral, true, [], insecureContentStatus, null));
    InspectorTest.addResult("\nSecurity Explanations (Mixed Form + Mixed Passive) --------------");
    dumpExplanations();

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests that the mixed-form explanation appears correctly.</p>
</body>
</html>
