third_party/WebKit/LayoutTests/http/tests/inspector/security/mixed-form-expected.txt

Tests that the mixed-form explanation appears correctly.


Security Explanations (Mixed Form Only) --------------
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Non-secure Form
        </DIV>
        <DIV >
The page includes a form with a non-secure "action" attribute.
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-secure >
    <DIV class=security-property security-property-secure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Secure Resources
        </DIV>
        <DIV >
All resources on this page are served securely.
        </DIV>
    </DIV>
</DIV>

Security Explanations (Mixed Form + Mixed Passive) --------------
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Non-secure Form
        </DIV>
        <DIV >
The page includes a form with a non-secure "action" attribute.
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Mixed Content
        </DIV>
        <DIV >
The site includes HTTP resources.
        </DIV>
        <DIV class=security-mixed-content >
Reload the page to record requests for HTTP resources.
        </DIV>
    </DIV>
</DIV>
